% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mjar.R
\name{getAbstracts}
\alias{getAbstracts}
\title{getAbstracts function}
\usage{
getAbstracts(links)
}
\arguments{
\item{links}{vector returned by getLinks() function.}
}
\description{
This function scrapes the abstracts of the papers matching the search input of the user.
}
\examples{
getAbstracts()
}

