% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mjar.R
\name{getLinks}
\alias{getLinks}
\title{getLinks function}
\usage{
getLinks(mjaSearch)
}
\arguments{
\item{mjaSearch}{HTML content of the search URL}
}
\description{
This function scrapes the URLs of the papers matching the search input of the user.
}
\examples{
getLinks()
}

