% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mjar.R
\name{getTitles}
\alias{getTitles}
\title{getTitles function}
\usage{
getTitles(mjaSearch)
}
\arguments{
\item{mjaSearch}{HTML content of the search URL}
}
\description{
This function scrapes the titles of the papers matching the search input of the user.
}
\examples{
getTitles()
}

