% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mjar.R
\name{search_MJA}
\alias{search_MJA}
\title{search_MJA function}
\usage{
search_MJA(userQuery)
}
\arguments{
\item{userQuery}{search keywords given by the user.}
}
\description{
This function will return the search result in the form of a matrix containing 3 columns - titles, links, abstracts of top 10 articles matching the user input.
}
\examples{
search_MJA()
}

